+++ azure/functions/extension/app_extension_base.py
@@ -1,35 +1,26 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
-from typing import NamedTuple, List
 import abc
-import os
 from logging import Logger
-from .extension_hook_meta import ExtensionHookMeta
 from .extension_meta import ExtensionMeta
 from .._abc import Context
 
 
-# Defines the life-cycle hooks we support in a single trigger
-class FuncExtensionHooks(NamedTuple):
-    after_function_load: List[ExtensionHookMeta] = []
-    before_invocation: List[ExtensionHookMeta] = []
-    after_invocation: List[ExtensionHookMeta] = []
-
-
-class FuncExtension(metaclass=ExtensionMeta):
+class AppExtensionBase(metaclass=ExtensionMeta):
     """An abstract class defines the life-cycle hooks which to be implemented
     by customer's extension.
 
     Everytime when a new extension is initialized in customer function scripts,
-    the ExtensionManager._func_exts field records the extension to this
-    specific function name.
+    the _app_exts field records the extension to this specific function name.
+    To access an implementation of specific trigger extension, use
+    _app_exts[i].<hook_name>.ext_impl
     """
 
     @abc.abstractmethod
-    def __init__(self, trigger_name: str):
+    def __init__(self, auto_enabled: bool = False):
         """Constructor for extension. This needs to be implemented and ensure
-        super().__init__(trigger_name) is called.
+        super().__init__() is called.
 
         The initializer serializes the extension to a tree. This speeds
         up the worker lookup and reduce the overhead on each invocation.
@@ -40,14 +31,14 @@ class FuncExtension(metaclass=ExtensionMeta):
         trigger_name: str
             The name of trigger the extension attaches to (e.g. HttpTrigger).
         """
-        ExtensionMeta.set_hooks_for_trigger(trigger_name, self)
+        ExtensionMeta.set_hooks_for_app(self)
 
     # DO NOT decorate this with @abc.abstratmethod
     # since implementation by subclass is not mandatory
-    def after_function_load(self, logger: Logger,
-                            function_name: str,
-                            function_directory: str,
-                            *args, **kwargs) -> None:
+    def after_function_load_global(self, logger: Logger,
+                                   function_name: str,
+                                   function_directory: str,
+                                   *args, **kwargs) -> None:
         """This hook will be called right after a customer's function is loaded
 
         Parameters
@@ -61,13 +52,12 @@ class FuncExtension(metaclass=ExtensionMeta):
             The path to customer's function directory
             (e.g. /home/site/wwwroot/HttpTrigger)
         """
-        pass
 
 
     # DO NOT decorate this with @abc.abstratmethod
     # since implementation by subclass is not mandatory
-    def before_invocation(self, logger: Logger, context: Context,
-                          *args, **kwargs) -> None:
+    def before_invocation_global(self, logger: Logger, context: Context,
+                                 *args, **kwargs) -> None:
         """This hook will be called right before customer's function
         is being executed.
 
@@ -84,8 +74,8 @@ class FuncExtension(metaclass=ExtensionMeta):
 
     # DO NOT decorate this with @abc.abstratmethod
     # since implementation by subclass is not mandatory
-    def after_invocation(self, logger: Logger, context: Context,
-                         *args, **kwargs) -> None:
+    def after_invocation_global(self, logger: Logger, context: Context,
+                                *args, **kwargs) -> None:
         """This hook will be called right after a customer's function
         is executed.
 
@@ -101,38 +91,13 @@ class FuncExtension(metaclass=ExtensionMeta):
         pass
 
     @classmethod
-    def register_to_function(cls, filename: str) -> 'FuncExtension':
+    def register_to_app(cls) -> 'AppExtension':
         """Register extension to a specific trigger. Derive trigger name from
         script filepath and AzureWebJobsScriptRoot environment variable.
-
-        Parameters
-        ----------
-        filename: str
-            The path to current trigger script. Usually, pass in __file__.
 
         Returns
         -------
         FuncExtension
             The extension or its subclass
         """
-        script_root = os.getenv('AzureWebJobsScriptRoot')
-        if script_root is None:
-            raise ValueError(
-                'AzureWebJobsScriptRoot environment variable is not defined. '
-                'Please ensure the extension is running in Azure Functions.'
-            )
-
-        try:
-            trigger_name = os.path.split(
-                os.path.relpath(
-                    os.path.abspath(filename),
-                    os.path.abspath(script_root)
-                )
-            )[0]
-        except IndexError:
-            raise ValueError(
-                'Failed to parse trigger name from filename. Please ensure '
-                '__file__ is passed into the filename argument'
-            )
-
-        return cls(trigger_name)
+        return cls()
