+++ azure/functions/extension/extension_scope.py
@@ -1,0 +1,10 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+from enum import Enum
+
+
+class ExtensionScope(Enum):
+    UNKNOWN = 0
+    APPLICATION = 1
+    FUNCTION = 2
